[MSDK-789] Upgrade MAX32680 (ME20) Makefile system (#51)
* Upgrade MAX32680 Makefiles (except for RV loader)

* Add project.mk to MAX32680 examples

* Enable littleFS for WearLeveling

* Resolve conflict between new Makefile system and gcc_riscv.mk

The new Makefile system was setting "LINKERFILE" for all projects.
However, gcc_riscv.mk wants to override the linkerfile for RISC-V
projects.  As a result, the new system was failing to link RISC-V
projects because it was pointing to the incorrect linkerfile.

The core Makefile has been updated to only set the default LINKERFILE
for Arm projects.  We could probably move this to gcc.mk at some point.

* Pass in PROJECT to host app from RV_Arm_Loader