Revert SBT rules back to using 'all'
'release' was missing the '-R .sig', so the generated .sbin files would
not work with the SBL.  The .sig section is _reserved_ in the linkerfile
on compilation, but not actually populated until the 'sign_app' tool
is called.

Adding '-R .sig' to the 'release' rule gets it working with the SBL, but the
.sig section will not be present for non-secure parts.  So it's not
suitable to add this option to 'release' in general.