Fix otac/otas firmware loader
The old firmware loader had some issues with Makefile recursion that
were very difficult to resolve.  Applying the same strategy as the RISCV
loader of building the external application's firmware locally in the
project, then linking it "in-place" resolved the issue.

The external firmware app will now be built in build/buildfw