Enable .DEFAULT_GOAL without touching gcc.mk
- Move .DEFAULT_GOAL section above libs.mk.  Library Makefiles were
the first goals to be added to the list, so the default goal was
always set to the peripheral drivers library

- Improve .DEFAULT_GOAL selection logic

- Move SBT-config.mk after .DEFAULT_GOAL selection.  It needs to know
what MAKECMDGOAL is.

- Use override operators on .DEFAULT_GOAL to maintain compatibility with
legacy Makefiles.  For some reason, legacy RISC-V makefiles really depend
on ".DEFAULT_GOAL :=" in gcc_riscv.mk.  Doing it this way means we don't
have to touch core gcc files for this feature.