Making TRACE portable.
This will allow us to build application with our without trace, without
experiencing link errors. We will still have an issue with code size if
the library is built with TRACE=2 and the application is built with
TRACE=0. We won't see the trace messages, but the code size will be
elevated.

In the case where we go from TRACE=0 to TRACE=2, we won't see the
print statements at all.