Use PROJ_LDFLAGS and PROJ_LIBS to link the library
Move myLib.a to lib/ and tell the linker the location of the library
using PROJ_LDFLAGS. Rename myLib.a to libmyLib.a so that it can be
passed to linker with the -l option.

Remove documentation for SBT option in Library_Generate project since it
does not output an executable that needs to be signed.