Add FaceID Examples for MAX78002, Add requirements.txt for db_gen Utility, Fix AttributeError for db_gen Utility (#537)
* ported faceID from MAX78000 and clang-formatted

* updated TFT folder name

* Removed unused MAX78000 images

* clean up and update to README

* Add requirements.txt for faceid_evkit db_gen

* Fix AttributeError (deprecated np.int for numpy > 1.20)

- AttributeError: module 'numpy' has no attribute 'int'.
`np.int` was a deprecated alias for the builtin `int`. To avoid this error in existing code, use `int` by itself. Doing this will not modify any behavior and is safe. When replacing `np.int`, you may wish to use e.g. `np.int64` or `np.int32` to specify the precision. If you wish to review your current use, check the release note link for additional information.
The aliases was originally deprecated in NumPy 1.20; for more details and guidance see the original release note at:
    https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations. Did you mean: 'inf'?

* Fix gen_db.sh script

* (MSDKGen) Update faceid_evkit project files

* Update readme with jumper settings

---------

Co-authored-by: Jake Carter <Jake.Carter@analog.com>