Compiling PAL with interrupt handlers.
This will cause the files that have interrupt handlers to be compiled
for every application instead of into the Cordio library. We need to
make sure that these objects are linked before the startup.S files
so the weak interrupt handlers in the C code is used instead of in the
startup file.