feat(Other): Enable building zephyr for risc-v targets
Configures the zephyr build system to use the appropriate cmsis system
startup and peripheral driver sources based on the target core
architecture. The mxc lock driver is not implemented for risc-v and
emits a build warning (which twister promotes to an error), therefore
the zephyr build system is configured to build it only for arm.

Before this patch, we could only build zephyr for arm targets, but now
we can build zephyr for arm or risc-v targets.

Signed-off-by: Maureen Helm <maureen.helm@analog.com>