Fix flash ecc issue on MAX32657
Erasing flash will also erase the ECC bits. These bits are not
updated until a flash write. Reading from erased memory will
signal a ECC error that is falsely corrected from 0xFF to 0xFD
on the 16th byte of each 128-bit line.

This commit applies a workaround by setting the 16th byte of each
line to 0xFF to handle this issue on max32657.

Co-authored-by: Mert Vatansever <mert.vatansever@analog.com>
Co-authored-by: Kevin Gillespie <kevin.gillespie@analog.com>