fix(Other): Fix Wrap_MXC_CAN_Init signature
The previous implementation declared Wrap_MXC_CAN_Init as void
and discarded the return value of the underlying MXC_CAN_Init
calls. This is incompatible with the can_max32 driver which
does eror checking and expects an int return value.

While this is a "breaking" change, no code depends on this
wrapper, because its signature is invalid.

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>