feat(CMSIS): CFSIO-4121 Call PeripheralInit
This commit adds a weakly defined function `PeripheralInit` for each SoC
which is called in `SystemInit` after `Board_Init` to initialize the
peripherals. This function is particular useful for tools which can
generate the `PeripheralInit` function based on a user's configuration
in a tool. The call to `PeripheralInit` is intentionally after
`Board_Init` so that the user defined peripheral initialization is not
overwritten.