Remove NOTE if PROJ_LDFLAGS looks ok
I didn't like the NOTE about PROJ_LDFLAGS coming out in the default build system when nothing is overridden.
So do a check for the very simplest case (making it fully checking is difficult and verbose) - that of
-Lc:/foo/bar. If PROJ_LDFLAGS is only -Lpath, and path contains a colon, and doesn't contain a backslash,
then assume we don't need the Note. Otherwise, in case of any doubt, emit the message.

Signed-off-by: Michael Perkins <michael.perkins@analog.com>