AFE: Handle Unexpected Startup Status
When afe_load_trims is called, AFE status in sys_ctrl is
expected to indicate a True POR, or a Reset has occurred.

In some cases the status will not be either of these.
Changed function to fully reinitialize afe state in this case.
Note, sometimes behavior seen when using debugger.